LayoutTests/platform/glib/accessibility/render-counter-text-expected.txt

This tests that the text for RenderCounter is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS line1.stringValue is 'AXValue: 1 : line1'
PASS line2.stringValue is 'AXValue: 2 : line2'
PASS successfullyParsed is true

TEST COMPLETE
